Add contribution guidelines with a git usage policy
Problem: It is important to maintain our history clean and
meaningful so that the review process is easy, the merge conflicts
are minimized, and the newcomers can easily understand how the
code has evolved.

Solution: Add contrinution guidelines, specifying one possible
flow we're going to use. This guide is supposed to be extended
with references to style guide, notes on maintaining the historical
versions of the contracts (G1, ...GN), etc. later.