Add the splitXVSRewards upgrade logic
Problem: We need to set supply speed = borrow speed = old speed
for all markets upon initializing the new version.

Solution: Add the corresponding function, call it from _become;
remove all entries from venusSpeeds[] and the array itself to
get some gas refund.

Note: Contrary to the corresponding [Compound's PR][comp144], we
do not need to check and set {supply,borrow}Index for each market,
as this code is related to the [markets initialization bug][bug],
which, although does affect Venus in some cases, hasn't been
triggered on Venus (because it so happened we always used
_setVenusSpeed in the same tx as _addMarkets). The bug itself has
been fixed in one of the previous commits.

[comp144]: https://github.com/compound-finance/compound-protocol/pull/144/files#diff-8ef9294768226293f278525b40998f11db7ee32464279f47ca7ad35294bb6a3fR1099
[bug]: https://www.comp.xyz/t/comptroller-compspeed-bug/2111