Fix venus{Supply,Borrow}State initialization bug
Problem: Currently, venus{Supply,Borrow}State structure is initialized
upon call to _setVenusSpeed. The logic is as follows: if the current
speed is zero and the new speed is nonzero, we check if both index
and block are nonzero. If this is the case, we consider the market
already initialized, and otherwise we initialize it. This leads to
the following bugs (originally [found][1] in Compound):

- If updateVenus{Supply,Borrow}Index gets called in between the calls
  to _supportMarket and _setVenusSpeed (after adding and before the
  initialization), the market would get bricked, as the former updates
  venus{Supply,Borrow}State.block, which makes _setVenusSpeed skip
  the initialization.
- Setting the speed to A -> 0 -> B (in separate blocks) would lead to
  extra interest accrued for the time the speed was zero.

[1]: https://www.comp.xyz/t/comptroller-compspeed-bug/2111

Solution:
- Add the initialization logic into a separate internal method
- Call this method from _supportMarket
- Remove the market initialization logic from _setVenusSpeed