Soften the requirements on gas limits
Problem: In certain conditions (when the market is updated for the
first time and the distribute{Borrower,Supplier}Venus is called for
the first time, the computation consumes more gas. This happens
because SSTORE now updates the values from zero to nonzero (init
index) rather than zero to zero or nonzero to nonzero (which both
are cheaper).

Solution: Soften the requirements on gas consumption.