Exclude "pending for withdrawal" tokens from deposit voting power
Problem: Users may put tokens in the vault just to participate in
voting in one proposal. The current implementation allows doing
this — they need to deposit and request the withdrawal immediately.
Taking the recent malicious proposals and the current unfair XVS
allocation (with several big whales) into account, the ability to
request withdrawals immediately and still have voting rights poses
a risk to our new governance model.

Solution: deny the "pending for withdrawal" tokens of the voting
rights by computing the voting power as a difference between the
staked amount and the pending withdrawals.