Add governance guardian to governor bravo
Problem: We need to retain the guardian capabilities after we migrate
to bravo. We need the right to veto the proposals because the current
XVS allocation allows malicious actors to take over the governance
— we need to mitigate this risk, although introducing some degree of
centralization.

Solution: Add governance guardian who can cancel the proposals but
CAN NOT change the governance contract.