Change governor bravo scenarios to use vault-based governance
Problem: The scenarios for GovernorBravo use XVS-based governance.
This is not what is going to happen in reality. We need to use
the vault instead of XVS.

Solution: Use the newly-added XVS Vault DSL constructs to deploy
the vault; use the vault as the source of getPriorVotes; update
block numbers to account for the vault deployment.