PVE003 Explicitly forbid deflationary tokens
Problem: The vault expects that the transferred amount of tokens
equals to the amount passed to transferFrom. In reality, though,
not all tokens act this way. Some of them take small fees for each
transfer.

Solution: Fixing this issue would increase the gas costs for
execution. Since we don't expect any tokens except XVS, VRT,
VAI to be used in the vault, we acknowledge this issue and
add a comment with a warning.