PVE004 Add pool validity checks
Problem: When a pool does not exist, we fail with a weird message
about boundaries.

Solution: Check the existece of the pool explicitly.

Implementation note: We use an internal function instead of the
modifier here to reduce the total contract size and to overcome
the compiler restrictions on the stack size in a least painful way.