Add Liquidator contract
Problem: We should split the liquidation incentive between
liquidators and treasury. To do this, we need to either upgrade all
vTokens (which is not possible due to vBNB), or come up with an
intermediary contract that pretends to be a liquidator, and then
splits the received vTokens in the right proportion.

Solution: This commit implements the second approach. Liquidator
contract is supposed to be the only address allowed to liquidate
borrows.