Add hardhat config and simulation helpers
Problem: Saddle is not always stable and sometimes works weirdly
with bsc. Moreover, the new developers coming to the project are
usually more comfortable with hardhat. We will test the VIPs using
hardhat, and we will also use hardhat to test the new features of
Venus.

Solution: Add a basic config and some helper scripts for VIP
simulations.