Add more test coverage to the voting power computations
Problem: We had a bug in the contract that resulted in an incorrect
voting power computation in case of multiple delegations to the same
address. We need to increase testing coverage to ensure that such
issues do not occur later.

Solution: Add tests for getPriorVotes — the main function used to
query the voting power. Although the function is tested in the XVS
contract, here we need to provide tests as well to ensure future-safety.