XVSVesting - 101 : copying the vestingsOfRecipient[i] in the storage to memory, which will cause performance loss.
Problem:
In the function getWithdrawableAmount(address recipient) view public nonZeroAddress(recipient),
the VestingRecord type variable vesting was storage in memory,
in fact this is equivalent to copying the vestingsOfRecipient[i] in the storage to memory, which will cause performance loss.

Solution:
Since in the function getWithdrawableAmount(address recipient)  the vesting variable won’t be modified, we recommend using storage rather than memory.