Remove VAI flywheel
This commit removes venusVAIRate public getters from all the
contracts (including the historical versions) to save up some
space and to get rid of obsolete code.

venusVAIRate hasn't been used for over a year, and VAI minter XVS
have been removed from the comptroller in the latest upgrade.

There is no point in keeping the obsolete code and tests in the
repo.

Historical versions (G*) are on their way to being removed as well,
so removing public getters from them is a step in this direction –
we want to make sure there are no tests that depend on this
functionality.

This change adds 25 bytes to our disposal.