Replace manual admin checks with ensureAdmin{,Or}
Problem: The contract is too close to the size limit, we need to
optimize aggressively.

Solution: Replace manual checks with function calls to ensureAdmin
(where only the admin can invoke the action) or ensureAdminOr(address)
(where either admin or a certain address can invoke the action).
Revert if the permission check fails.

This optimization saves us 512 bytes of contract size.

This is a breaking change.