Add fine grained pause control
Problem: In case of any incident, we would like to be able to just
pause certain actions rather than the whole protocol.

Solution: Add _setActionsPaused function that accepts a list of
markets and actions to pause on these markets. Refactor the current
approach to pausing to save some Comptroller space.

Note: This change is a bit controversial since it does allow pause
guardians to temporarily disable repayBorrow and redeem. This might
be worrisome for the community, so I propose we discuss this change
publicly before upgrading.