Add missing implementations to BoolComptroller
Problem: Commit c8eaaebb20fd57b2398506b9416ebf3d40bbf603 added more
methods to ComptrollerInterface so that it could be used in other
contracts instead of Comptroller directly. However, this introduced
a problem, because BoolComptroller in tests inherited from
ComptrollerInterface. Some of the methods lacked implementation,
but since we're on an ancient compiler version, this lead to a
deploy-time failure instead of the compile time failure.

Solution: Add stub implementations that just revert, so that the
deployment of BoolComptroller succeeds.