Use underlying supply in supply cap
Problem: The code in Comptroller uses totalSupply (i.e. vToken
supply) rather than the underlying supply for the cap.
Using underlying is more correct here: we can't add totalSupply
and mintAmount, because the former is in vTokens while the latter
is in underlying.

Solution: Compute underlying supply as vToken supply * exchange rate