tests: refactor XVSVault test suite, make tests more atomic
Problem: The existing test suite for XVSVault is written in
"scenario" manner – instead of testing the behavior of individual
functions it runs a series of state updates and assertions, joined
by a single real-world use-case. While this is useful to ensure
the vault behaves as expected in general, it makes it difficult
to add new atomic, function-oriented checks to the suite.

Solution: Refactor the test suite, adding atomic checks to each
function previously tested in "scenario" manner. Move the most
important and complex scenario to a separate "Scenarios" describe
block.