fix: forbid zero alloc points in XVSVault
Problem: If a pool is added with alloc points equal to zero, and it
is the only pool per reward token, total alloc points would be zero,
leading to division by zero error in massUpdatePools. It is not
possible to recover by calling set() with a different alloc points
parameter because set() calls massUpdatePools under the hood. This
can lead to denial of service because deposits, withdrawals and
claims will not be possible either. Misconfiguring a pool using
set() call can have the same impact if there is no other pool for
this reward token.

Solution:
1. Forbid adding a pool with zero alloc points. This should never
happen because it does not make sense from the business point of
view.
2. ALLOW setting zero alloc points for existing pools IF there is
another pool with nonzero alloc points configured for this reward
token; forbid otherwise. Although it is an unlikely use case, it's
still valid from the business perspective, e.g. if we decide to
discontinue rewarding XVS stakers with XVS but, e.g., reward VAI
stakers with XVS instead.