refactor: move typedefs out of ComptrollerV*Storage
Problem: When we import enums, we don't usually care at which
storage version this enum was introduced. Moreover, enums can be
extended with time, and this does not pose any risk of storage
collision. Having ComptrollerV*Storage in interface imports is
misleading and hard to maintain.

Solution: Move Action enum (and possibly other enums in future) to
a separate ComptrollerTypes interface. Ideally, this could be a
free enum, but solc 0.5 does not allow enums outside of interfaces.